changes to use references
I added some references for the long and short name of the iTC and then used those in the text. I wil update other templates similarly as I think this will make things a lot easier in the future.

I also removed the TOC since I don't think that is needed for this as it would probably just be submitted as HTML to NIAP to post.