Update to the out of scope descriptive note
In response to @researchcollab. I had looked at the ND ESR and came up with the text from that. I have moved that to the end as an example of things that may be excluded but could become later ESRs and provided an example such as the USB one for the first part.