Asciidoctor extension for custom admonitions
This is the content needed to create custom admonitions when building HTML files from adoc.

This is for #38.

This will have to be installed on any system generating files for publication so these icons will show up properly for everyone. The instructions will be updated in the guidance.

PDF version of the same will be publised separately.