Admin site
This is my version of the admin site.

I built from what @atsecsysad had uploaded (if you look at the various files you will see direct copies in several of them).

What I am going for here as I kept looking at the admin docs was that the entire set could be done as more of a website instead of a few PDF files. So what I did was to create a few different pages along with a main one, and then point to references in a few of the other documents.