Updates based on feedback
@ansukert I have split the original document into 2 and then added an email template (in text instead of adoc).

The existing template is now more focused on the actual announcement and details about the document under review.

A new document was created specifically about the review process, showing the timelines and such but nothing about docs under review. This page does have the comment spreadsheet though.

The email template is a blending of the PWG one and one I had put together for the Bio-iTC. I tried to provide the framework for it while providing a simple example of a ballot measure. I have assumed the actual ballot measure would be explicitly laid out in the email, not something where you are told to read something else to make the decision (i.e. the question is in another document, not the case where you are approving another document).