Update to guide with links to GitHub help
I added links to GitHub Help pages for the major sections (and some minor ones). Also added a note that this is in there (to search for GitHub Help).

Also removed the resources section since I expect this to be a web document instead of a PDF, so everything will be corss linked in some way.