Admonition and dummy iTC update
This update covers the dummy iTC updates as well as bringing in the new admonition texts.

Check https://itc-wgtools.github.io/templates/cPP-template.html to see the output with the icons that will be generated in the output to mark things. Note that these won't display when viewing in GitHub or generally in most editors (there may be a way to make it show up that is a little more complicated that I haven't tried yet), but the main point was to show the admonitions with different markers to call out what they are, and then to have them show up in the output as "wrong" if you are using the standard ones so they could be easily picked out.