Updates related to template repositories
This is a set of updates to the web pages related to using the template repositories to setup the iTC workspace. This is an adjustment from the previous which had blank (i.e. they create a blank set).

I have also added a new page that talks about when to create these repositories (since they don't NEED to be done all at the same time).