Updates for Maintenance/IT
This update covers the changes that should be needed to cover setting up the organization and the repositories to also cover the needs of maintenance and IT.

Summary is creation of Editor and IT teams and then establishing permissions for them on the appropriate branches.

There are some other content changes based on better understandings.